extract prepareRapidastConfig to separate file
